+++ R/json_structure.R
@@ -27,12 +27,12 @@
 #'         represented as <parent>.<index> where <parent> is the ID for the
 #'         parent and <index> is this index.
 #'
-#'   \item \code{seq} the sequence of keys / indices that led to this child
+#'   \item \code{seq} the sequence of names / indices that led to this child
 #'         (parents that are arrays are excluded) as a list, where character
 #'         strings denote objects and integers denote array positions
 #'
-#'   \item \code{key} if this is the value of an object, what was the key that
-#'         it is listed under (from \code{\link{gather_keys}}).
+#'   \item \code{name} if this is the value of an object, what was the name that
+#'         it is listed under (from \code{\link{gather_object}}).
 #'
 #'   \item \code{type} the type of this object (from \code{\link{json_types}}).
 #'
@@ -52,7 +52,7 @@
 #' '"string"' %>% json_structure
 #'
 #' # A simple object
-#' '{"key": "value"}' %>% json_structure
+#' '{"name": "value"}' %>% json_structure
 #'
 #' # A complex array
 #' '[{"a": 1}, [1, 2], "a", 1, true, null]' %>% json_structure
@@ -92,7 +92,7 @@ json_structure_init <- function(x) {
       index = 1L,
       child.id = "1",
       seq = replicate(n(), list()),
-      key = NA_character_
+      name = NA_character_
     ) %>%
     json_types %>%
     json_lengths
@@ -119,7 +119,7 @@ json_structure_empty <- function() {
       index = integer(0),
       child.id = character(0),
       seq = list(),
-      key = character(0),
+      name = character(0),
       type = factor(character(0), levels = allowed_json_types),
       length = integer(0)
     ),
@@ -162,7 +162,7 @@ json_structure_objects <- function(s) {
       seq,
       level = level + 1L
     ) %>%
-    gather_keys %>%
+    gather_object %>%
     json_types %>%
     json_lengths
 
@@ -173,10 +173,10 @@ json_structure_objects <- function(s) {
     ungroup %>%
     mutate(
       child.id = paste(parent.id, index, sep = "."),
-      seq = map2(seq, key, c)
+      seq = map2(seq, name, c)
     ) %>%
     select(
-      document.id, parent.id, level, index, child.id, seq, key, type, length
+      document.id, parent.id, level, index, child.id, seq, name, type, length
     )
 
   # Reconstruct tbl_json object
@@ -203,7 +203,7 @@ json_structure_arrays <- function(s) {
     ) %>%
     transmute(
       document.id, parent.id, level, index, child.id,
-      seq, key = NA_character_, type, length
+      seq, name = NA_character_, type, length
     )
 
 }