+++ tests/testthat/test-append_values.R
@@ -1,29 +1,29 @@
 context("append_values")
 
 test_that("has correct complete structure with simple input", {
-    
+
     json <- '[{"name": "anne"}, {"name": "bob"}, {"name": "george"}]'
 
     expect_identical(
-      json %>% gather_array %>% gather_keys %>% 
+      json %>% gather_array %>% gather_object %>%
         append_values_string,
       tbl_json(
         data.frame(
           document.id = c(1L, 1L, 1L),
           array.index = 1L:3L,
-          key = rep("name", 3),
+          name = rep("name", 3),
           string = c("anne", "bob", "george"),
           stringsAsFactors = FALSE
         ),
         list("anne", "bob", "george")
       )
     )
-    
+
   }
 )
 
 test_that("string works with value array", {
-    
+
     json <- '["a", "b"]'
 
     expect_identical(
@@ -38,49 +38,49 @@ test_that("string works with value array", {
         list("a", "b")
       )
     )
-    
+
   }
 )
 
 test_that("string works with simple input", {
-    
+
     json <- '["a", "b", null]'
 
     expect_identical(
       (json %>% gather_array %>% append_values_string)$string,
       c("a", "b", NA_character_)
     )
-    
+
   }
 )
 
 test_that("number works with simple input", {
-    
+
     json <- '[1, 2, null]'
 
     expect_identical(
       (json %>% gather_array %>% append_values_number)$number,
       c(1, 2, NA_real_)
     )
-    
+
   }
 )
 
 test_that("logical works with simple input", {
-    
+
     json <- '[true, false, null]'
 
     expect_identical(
       (json %>% gather_array %>% append_values_logical)$logical,
       c(TRUE, FALSE, NA)
     )
-    
+
   }
 )
 
 test_that("handles mixed input as appropriate NA", {
-    
-    data <- '["a", 1, true, null]' %>% gather_array 
+
+    data <- '["a", 1, true, null]' %>% gather_array
     expect_identical(
       (data %>% append_values_string)$string,
       c("a", "1", "TRUE", NA_character_)
@@ -96,15 +96,15 @@ test_that("handles mixed input as appropriate NA", {
       (data %>% append_values_logical)$logical,
       c(NA, NA, TRUE, NA)
     )
-    
+
   }
 )
 
 test_that("correctly handles character(0)", {
-    
+
     empty <- tbl_json(
       data.frame(
-        document.id = integer(0), 
+        document.id = integer(0),
         string = character(0),
         stringsAsFactors = FALSE),
 
@@ -118,13 +118,13 @@ test_that("correctly handles character(0)", {
 )
 
 test_that("correctly handles {}", {
-    
+
     nl <- list()
     names(nl) <- character(0)
-    
+
     empty <- tbl_json(
       data.frame(
-        document.id = 1L, 
+        document.id = 1L,
         string = NA_character_,
         stringsAsFactors = FALSE),
       list(nl))
@@ -132,15 +132,15 @@ test_that("correctly handles {}", {
     expect_identical(
       '{}' %>% append_values_string,
       empty)
-    
+
   }
 )
 
 test_that("correctly handles []", {
-    
+
     empty <- tbl_json(
       data.frame(
-        document.id = 1L, 
+        document.id = 1L,
         string = NA_character_,
         stringsAsFactors = FALSE),
       list(list()))
@@ -148,13 +148,13 @@ test_that("correctly handles []", {
     expect_identical(
       '[]' %>% append_values_string,
       empty)
-    
+
   }
 )
 
 test_that("correctly handles mixed types when force=FALSE", {
 
-    data <- '["a", 1, true, null]' %>% gather_array 
+    data <- '["a", 1, true, null]' %>% gather_array
 
     expect_identical(
       (data %>% append_values_string(force=FALSE))$string,
@@ -174,49 +174,49 @@ test_that("correctly handles mixed types when force=FALSE", {
 test_that("correctly handles append when trying to append an array", {
 
    data <- '[["a", "b", "c"], "d", "e", "f"]' %>% gather_array
-   
+
    expect_identical(
      (data %>% append_values_string())$string,
      c(NA_character_, "d", "e", "f")
-   ) 
+   )
   }
 )
 
 test_that("recursive works as expected", {
 
-   data <- '{"item1": {"price" : 30}, "item2" : 40, "item3" : 30}' %>% gather_keys 
+   data <- '{"item1": {"price" : 30}, "item2" : 40, "item3" : 30}' %>% gather_object
    expected_na <- c(NA_real_, 40, 30)
    expected_val <- c(30, 40, 30)
 
    expect_identical(
-     (data %>% append_values_number(force=TRUE, recursive=FALSE))$number, 
+     (data %>% append_values_number(force=TRUE, recursive=FALSE))$number,
      expected_na)
    expect_identical(
-     (data %>% append_values_number(force=TRUE, recursive=TRUE))$number, 
+     (data %>% append_values_number(force=TRUE, recursive=TRUE))$number,
      expected_val)
    expect_identical(
-     (data %>% append_values_number(force=FALSE, recursive=FALSE))$number, 
+     (data %>% append_values_number(force=FALSE, recursive=FALSE))$number,
      expected_na)
    expect_error(
      (data %>% append_values_number(force=FALSE, recursive=TRUE))$number)
 
-   data <- '{"item1": {"price" : {"usd" : {"real" : 30}}}, "item2" : 40, "item3" : 30}' %>% 
-              gather_keys 
+   data <- '{"item1": {"price" : {"usd" : {"real" : 30}}}, "item2" : 40, "item3" : 30}' %>%
+              gather_object
 
    expect_identical(
-     (data %>% append_values_number(recursive=FALSE))$number, 
+     (data %>% append_values_number(recursive=FALSE))$number,
      expected_na)
    expect_identical(
-     (data %>% append_values_number(recursive=TRUE))$number, 
+     (data %>% append_values_number(recursive=TRUE))$number,
      expected_val)
 
-   data <- '{"item1": {"price" : 30, "qty" : 1}, "item2" : 40, "item3" : 30}' %>% gather_keys 
+   data <- '{"item1": {"price" : 30, "qty" : 1}, "item2" : 40, "item3" : 30}' %>% gather_object
 
    expect_identical(
-     (data %>% append_values_number(recursive=FALSE))$number, 
+     (data %>% append_values_number(recursive=FALSE))$number,
      expected_na)
    expect_identical(
-     (data %>% append_values_number(recursive=TRUE))$number, 
+     (data %>% append_values_number(recursive=TRUE))$number,
      expected_na)
 
   }