+++ NEWS.md
@@ -16,6 +16,8 @@
 
 * `as.character.tbl_json()` converts `tbl_json` objects back into JSON character strings (#62)
 
+* `gather_object()` replaces `gather_keys()`, with default `column.name` of `name` instead of `key` (#66). This more closely matches the [JSON standard](http://www.json.org/), which refers to objects as name-value pairs, and is now consistent with `gather_array()`.
+
 ## Documentation changes
 
 * "Visualizing JSON" vignette for visualizing the structure of complex JSON data, like the `companies` example (#4)
@@ -28,7 +30,7 @@
 
 * Use [purrr](https://github.com/jeremystan/purrr) for most list based internal operations (#1)
 
-* Use [tidyr](https://github.com/hadley/tidyr) for `gather_array` and `gather_keys` functions (#28)
+* Use [tidyr](https://github.com/hadley/tidyr) for `gather_array` and `gather_object` functions (#28)
 
 * Imported the magrittr `%>%` operator (#17)
 
@@ -41,3 +43,7 @@
 ## Other changes
 
 * Migrated development to [jeremystan](https://github.com/jeremystan/tidyjson) from [sailthru](https://github.com/sailthru/tidyjson)
+
+## Deprecated functions
+
+* `gather_keys()` -> use `gather_object()` instead